Fetch subjects on demand Paginate subject URLs, from the data exports, then use computed data to fetch subjects individually as subject pages are built. Remove the subjectCollections build. Build all the subject pages together, using shared data.
