+++ package.json
@@ -18,7 +18,6 @@
     "build:tags": "NODE_ENV=production node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --input=src/tags",
     "build:site": "NODE_ENV=production node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js",
     "build:subjects-api": "NODE_ENV=production node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --input=src/subjects:api",
-    "build:subjectCollections": "NODE_ENV=production node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --input=src/subjectCollections",
     "build:subjects": "NODE_ENV=production node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --input=src/subjects",
     "clean": "rm -rf ./dist && mkdir -p ./dist/logs",
     "copy-logs": "mv build.log dist/logs/build.log",
@@ -29,7 +28,6 @@
     "dev:site": "NODE_ENV=development node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --watch",
     "dev:tags": "NODE_ENV=development node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --input=src/tags --watch",
     "dev:subjects-api": "NODE_ENV=development node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --input=src/subjects:api --watch",
-    "dev:subjectCollections": "NODE_ENV=development node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --input=src/subjectCollections --watch",
     "dev:subjects": "NODE_ENV=development node --max-old-space-size=${NODE_RAM:-12000} ./node_modules/@11ty/eleventy/cmd.js --input=src/subjects --watch",
     "verify": "node verify.js",
     "webpack:build:assets": "NODE_ENV=production webpack --mode=production",
