--- src/subjectCollections/_data/project.js
@@ -1,6 +1,0 @@
-const awaitProject = require('../../helpers/project');
-
-module.exports = async function () {
-  const project = await awaitProject;
-  return project;
-}
