--- src/subjectCollections/_includes/layouts/default.njk
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    {% pageMetadata page=page, title=title, description=description, ogImage=ogImage %}
-    <link rel="stylesheet" href="{% webpackAsset 'main.css' %}">
-  </head>
-  <body>
-    {% pageHeader %}
-    <div  class="container mx-auto">
-      {{ content | safe }}
-    </div>
-    <script src="{% webpackAsset 'main.js' %}"></script>
-  </body>
-</html>
