+++ src/subjects/_data/subjectURLs.js
@@ -1,0 +1,6 @@
+const awaitSubjectURLs = require('../../helpers/subjectURLs');
+
+module.exports = async function () {
+  const subjectURLs = await awaitSubjectURLs;
+  return subjectURLs;
+}
