--- src/subjects/_data/subjects.js
@@ -1,6 +1,0 @@
-const awaitSubjects = require('../../helpers/subjects');
-
-module.exports = async function () {
-  const subjects = await awaitSubjects;
-  return subjects;
-}
