+++ src/subjects/subjects/subject.11tydata.js
@@ -1,7 +1,10 @@
+const API = require('../../helpers/api');
+
 module.exports = {
   eleventyComputed: {
+    subject: async data => await API.get(data.subjectURL),
+    ogImage: data => data.subject.location,
     title: data => `Subject ${data.subject.zooniverse_id}`,
-    description: data => `Subject ${ data.subject.zooniverse_id } from ${ data.project.display_name }.`,
-    ogImage: data => data.subject.location
+    description: data => `Subject ${ data.subject.zooniverse_id } from ${ data.project.display_name }.`
   }
 }
