+++ src/subjects:api/subjects/api.11tydata.js
@@ -1,0 +1,7 @@
+const API = require('../../helpers/api');
+
+module.exports = {
+  eleventyComputed: {
+    subject: async data => await API.get(data.subjectURL)
+  }
+}
